configuração inicial das pastas
