--- src/main/java/com/maua/roubit/dto/LoginRequestDTO.java
@@ -1,4 +1,0 @@
-package com.maua.roubit.dto;
-
-public record LoginRequestDTO(String email, String password) {
-}
