--- src/main/java/com/maua/roubit/dto/RegisterRequestDTO.java
@@ -1,4 +1,0 @@
-package com.maua.roubit.dto;
-
-public record RegisterRequestDTO(String name, String email, String password) {
-}
