--- src/main/java/com/maua/roubit/dto/ResponseDTO.java
@@ -1,4 +1,0 @@
-package com.maua.roubit.dto;
-
-public record ResponseDTO(String name, String token) {
-}
