+++ src/main/java/com/maua/roubit/modules/User/authUser/app/AuthUserController.java
@@ -1,0 +1,54 @@
+package com.maua.roubit.modules.User.authUser.app;
+
+import com.maua.roubit.shared.domain.entities.User;
+import com.maua.roubit.shared.domain.dto.LoginRequestDTO;
+import com.maua.roubit.shared.domain.dto.RegisterRequestDTO;
+import com.maua.roubit.shared.domain.dto.ResponseDTO;
+import com.maua.roubit.shared.infra.security.TokenService;
+import com.maua.roubit.shared.domain.repositories.UserRepo;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Optional;
+
+@RestController
+@RequestMapping("/auth")
+@RequiredArgsConstructor
+public class AuthUserController {
+    private final UserRepo repository;
+    private final PasswordEncoder passwordEncoder;
+    private final TokenService tokenService;
+
+    @PostMapping("/login")
+    public ResponseEntity login(@RequestBody LoginRequestDTO body){
+        User user = this.repository.findByEmail(body.email()).orElseThrow(() -> new RuntimeException("User not found"));
+        if(passwordEncoder.matches(body.password(), user.getPassword())) {
+            String token = this.tokenService.generateToken(user);
+            return ResponseEntity.ok(new ResponseDTO(user.getName(), token));
+        }
+        return ResponseEntity.badRequest().build();
+    }
+
+
+    @PostMapping("/register")
+    public ResponseEntity register(@RequestBody RegisterRequestDTO body){
+        Optional<User> user = this.repository.findByEmail(body.email());
+
+        if(user.isEmpty()) {
+            User newUser = new User();
+            newUser.setPassword(passwordEncoder.encode(body.password()));
+            newUser.setEmail(body.email());
+            newUser.setName(body.name());
+            this.repository.save(newUser);
+
+            String token = this.tokenService.generateToken(newUser);
+            return ResponseEntity.ok(new ResponseDTO(newUser.getName(), token));
+        }
+        return ResponseEntity.badRequest().build();
+    }
+}
