+++ src/main/java/com/maua/roubit/modules/User/authUser/app/AuthUserViewmodel.java
@@ -1,0 +1,4 @@
+package com.maua.roubit.modules.User.authUser.app;
+
+public class AuthUserViewmodel {
+}
