+++ src/main/java/com/maua/roubit/modules/User/createUser/app/CreateUserUsecase.java
@@ -1,0 +1,4 @@
+package com.maua.roubit.modules.User.createUser.app;
+
+public class CreateUserUsecase {
+}
