+++ src/main/java/com/maua/roubit/modules/User/deleteUser/app/DeleteUserViewmodel.java
@@ -1,0 +1,4 @@
+package com.maua.roubit.modules.User.deleteUser.app;
+
+public class DeleteUserViewmodel {
+}
