+++ src/main/java/com/maua/roubit/modules/User/getAllUsers/app/GetAllUsersViewmodel.java
@@ -1,0 +1,4 @@
+package com.maua.roubit.modules.User.getAllUsers.app;
+
+public class GetAllUsersViewmodel {
+}
