+++ src/main/java/com/maua/roubit/modules/User/getUser/app/GetUserUsecase.java
@@ -1,0 +1,4 @@
+package com.maua.roubit.modules.User.getUser.app;
+
+public class GetUserUsecase {
+}
