+++ src/main/java/com/maua/roubit/modules/User/getUser/app/GetUserViewmodel.java
@@ -1,0 +1,4 @@
+package com.maua.roubit.modules.User.getUser.app;
+
+public class GetUserViewmodel {
+}
