--- src/main/java/com/maua/roubit/repository/UserRepo.java
@@ -1,10 +1,0 @@
-package com.maua.roubit.repository;
-
-import com.maua.roubit.domain.user.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepo extends JpaRepository<User, String> {
-    Optional<User> findByEmail(String email);
-}
