+++ src/main/java/com/maua/roubit/shared/domain/dto/LoginRequestDTO.java
@@ -1,0 +1,4 @@
+package com.maua.roubit.shared.domain.dto;
+
+public record LoginRequestDTO(String email, String password) {
+}
