+++ src/main/java/com/maua/roubit/shared/domain/dto/RegisterRequestDTO.java
@@ -1,0 +1,4 @@
+package com.maua.roubit.shared.domain.dto;
+
+public record RegisterRequestDTO(String name, String email, String password) {
+}
