+++ src/main/java/com/maua/roubit/shared/domain/dto/ResponseDTO.java
@@ -1,0 +1,4 @@
+package com.maua.roubit.shared.domain.dto;
+
+public record ResponseDTO(String name, String token) {
+}
