+++ src/main/java/com/maua/roubit/shared/domain/repositories/UserRepo.java
@@ -1,0 +1,10 @@
+package com.maua.roubit.shared.domain.repositories;
+
+import com.maua.roubit.shared.domain.entities.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepo extends JpaRepository<User, String> {
+    Optional<User> findByEmail(String email);
+}
