+++ src/main/java/com/maua/roubit/shared/infra/security/CustomUserDetailsService.java
@@ -1,0 +1,23 @@
+package com.maua.roubit.shared.infra.security;
+
+import com.maua.roubit.shared.domain.entities.User;
+import com.maua.roubit.shared.domain.repositories.UserRepo;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Component;
+
+import java.util.ArrayList;
+
+@Component
+public class CustomUserDetailsService implements UserDetailsService {
+    @Autowired
+    private UserRepo repository;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        User user = this.repository.findByEmail(username).orElseThrow(() -> new UsernameNotFoundException("User Not Found"));
+        return new org.springframework.security.core.userdetails.User(user.getEmail(), user.getPassword(), new ArrayList<>());
+    }
+}
