+++ src/main/java/com/maua/roubit/shared/infra/security/cors/CorsConfig.java
@@ -1,0 +1,16 @@
+package com.maua.roubit.shared.infra.security.cors;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class CorsConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins("http://localhost:4200")
+                .allowedMethods("GET", "POST", "DELETE", "PUT");
+    }
+}
