+++ src/test/java/com/maua/roubit/modules/User/getAllUsers/app/GetAllUsersControllerTests.java
@@ -1,0 +1,4 @@
+package com.maua.roubit.modules.User.getAllUsers.app;
+
+public class GetAllUsersControllerTests {
+}
