+++ src/test/java/com/maua/roubit/shared/entities/UserTests.java
@@ -1,0 +1,2 @@
+package com.maua.roubit.shared.entities;public class UserTests {
+}
