+++ src/test/java/com/maua/roubit/shared/infra/dto/UserDBTests.java
@@ -1,0 +1,4 @@
+package com.maua.roubit.shared.infra.dto;
+
+public class UserDBTests {
+}
