+++ src/test/java/com/maua/roubit/shared/infra/repositories/UserRepository/UserRepositoryMockTests.java
@@ -1,0 +1,4 @@
+package com.maua.roubit.shared.infra.repositories.UserRepository;
+
+public class UserRepositoryMockTests {
+}
